Use prelude_log for debugging.
Remove the per client mutex which was only used to prevent a race between server_logic_notify_write()
and remove_connection(). They are both always called from the same thread.

Make the close() connection callback return a value. If the return is -1, we assume that the close()
didn't complete because we hit EAGAIN. In this case, the connection is not removed from the set. 
server_logic_remove_client() now return -1 on EAGAIN.

Cleanup in handle_fd_event(). Separate the case where we saw no event with the case where an error
was returned. No need to check for POLLHUP|POLLERR|POLLNVAL if the previous call returned an error
or asked to stop processing.


git-svn-id: file:///home/yoann/SVN/trunk/prelude-manager@5281 09c5ec92-17d4-0310-903a-819935f44dba