More improvement to the poll module bug fix commited previously. Here is the full log:
There is a problem where GnuLib replacement poll() would fail with a 
non connected socket (example: server socket polled for incoming connection 
event).

This patch fix the issue in the following way:
- [ALL except OSX]: if recv() return -1 with errno set to ENOTCONN, then
trigger a POLLIN.

- [OSX]: The situation is more complicated since recv() was not used due
to a OSX bug where data would be consumed despite the MSG_PEEK flag
being set. FIONREAD was used in place of the recv() call.

However, FIONREAD does not permit to determine whether the connection
has been closed, or if we are working with a non connected socket (it
will return 0/set available data to 0 in both case). 

Instead of using FIONREAD, modify the check to use recv() with a 0
length (which is, AFAIK, not portable, but seems to work under OSX).
This doesn't consume any data, but permit to make the distinction
between both case.


git-svn-id: file:///home/yoann/SVN/trunk/prelude-manager@8674 09c5ec92-17d4-0310-903a-819935f44dba