Make sure we listen on all interface (IPv4+IPv6 dual stack fixes)
There is a problem on Linux system where getaddrinfo() return address in
the wrong sort order (IPv4 first, IPv6 next):

As a result we used to first bind IPv4 addresses, but then we were getting
an error for dual-stacked addresses, when the IPv6 addresses came second.
When an address is dual-stacked, we thus ended-up listening only to the IPv4
instance.

The error happened on dual-stack Linux system, because mapping the IPv6
address will actually attempt to bind both the IPv4 and IPv6 address.

In order to prevent this problem, set the IPV6_V6ONLY option so that only
the IPv6 address is bound.

Additionally, The prelude-admin server code used to report success for
binding an interface even thought the operation failed. We now only report
successful bind.

Change-Id: I4d0fc89dc719a050de65cd16ed763c8826004473